# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Gloo",
        "title": "Gloo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gloo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "gloo-api-production.up.railway.app",
    "basePath": "/api",
    "paths": {
...
        "/search/{userId}": {
            "get": {
                "description": "Get all searches",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Get all searches",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved searches",
                        "schema": {
                            "$ref": "#/definitions/models.GetAllSearchesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/search/{userId}/suggestions": {
            "get": {
                "description": "Get all suggestions",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Get all suggestions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved searches",
                        "schema": {
                            "$ref": "#/definitions/models.GetAllSearchesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "put": {
                "description": "Update user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated user",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/change": {
            "get": {
                "description": "Get change",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get change",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved change",
                        "schema": {
                            "$ref": "#/definitions/models.GetChangeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/profile": {
            "get": {
                "description": "Get profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved profile",
                        "schema": {
                            "$ref": "#/definitions/models.GetProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}": {
            "get": {
                "description": "Get user by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Get user by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved user",
                        "schema": {
                            "$ref": "#/definitions/models.GetUserByIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ErrorBadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.ErrorInternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.ErrorNotFoundResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Not Found"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 404
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.GetAllSearchesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "created_at": {
                                "type": "string",
                                "example": "2021-01-01T00:00:00Z"
                            },
                            "created_by": {
                                "type": "string",
                                "example": "@facu.potti"
                            },
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "search_term": {
                                "type": "string",
                                "example": "Salty"
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved searches"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.GetChangeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "estimated_time": {
                                "type": "string",
                                "example": "25 minutes"
                            },
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "image_url": {
                                "type": "string",
                                "example": "https://example.com/image.png"
                            },
                            "title": {
                                "type": "string",
                                "example": "Chicken Wings"
                            },
                            "total_stars": {
                                "type": "number",
                                "example": 4.5
                            }
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved change"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.GetProfileResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "description": {
                            "type": "string",
                            "example": "I'm a software engineer"
                        },
                        "email": {
                            "type": "string",
                            "example": "facu.potti@gmail.com"
                        },
                        "id": {
                            "type": "integer",
                            "example": 1
                        },
                        "name": {
                            "type": "string",
                            "example": "Facu Potti"
                        },
                        "profile_picture": {
                            "type": "string",
                            "example": "https://example.com/image.png"
                        },
                        "total_followers": {
                            "type": "integer",
                            "example": 250
                        },
                        "total_following": {
                            "type": "integer",
                            "example": 120
                        },
                        "total_posts": {
                            "type": "integer",
                            "example": 120
                        },
                        "username": {
                            "type": "string",
                            "example": "@facu.potti"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved profile"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.GetUserByIdResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "properties": {
                        "description": {
                            "type": "string",
                            "example": "I'm a software engineer"
                        },
                        "email": {
                            "type": "string",
                            "example": "facu.potti@gmail.com"
                        },
                        "id": {
                            "type": "integer",
                            "example": 1
                        },
                        "name": {
                            "type": "string",
                            "example": "Facu Potti"
                        },
                        "profile_picture": {
                            "type": "string",
                            "example": "https://example.com/image.png"
                        },
                        "social_media": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "example": "https://twitter.com/facu.potti"
                                    }
                                }
                            }
                        },
                        "username": {
                            "type": "string",
                            "example": "@facu.potti"
                        }
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully retrieved user"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.UpdateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Updated successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.UpdateUserRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "I'm a software engineer"
                },
                "email": {
                    "type": "string",
                    "example": "facu.potti@gmail.com"
                },
                "name": {
                    "type": "string",
                    "example": "Facu Potti"
                },
                "profile_picture": {
                    "type": "string",
                    "example": "https://example.com/image.png"
                },
                "social_media": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "url": {
                                "type": "string",
                                "example": "https://twitter.com/facu.potti"
                            }
                        }
                    }
                },
                "username": {
                    "type": "string",
                    "example": "@facu.potti"
                }
            }
        }
    }
}